fix OomDetector::high_thresh_at_ resets
Summary:
OomDetector::high_thresh_at_ is used to determine how long the
pressure has stayed over the high threshold.  It currently is only
reset when high threshold OOM condition is detected, which leads to
the following problems.

* It doesn't get reset when the pressure drops below the threshold and
  two peaks with a period of low pressure can incorrectly trigger the
  verdict.

* It gets reset once the verdict is triggered and can only be
  triggered after another full threshold duration has passed.  The
  correct behavior would be keeping giving the verdict as long as
  pressure stays above the threshold.

This commit fixes the above issues by resetting
OomDetector::high_thresh_at_ when the observed pressure is below the
threshold and nowhere else.

Reviewed By: danobi

Differential Revision: D9222667

fbshipit-source-id: 1698d3715d7bead9f3b33aa1a5153977e6f74f70